Ensure city's shield-store cannot become negative
Changed functionality of 'removing added shields when the queue is empty' to not 'adding shields in case queue is empty'. The number of shields that were removed was bigger than the number of shields added which may result in negative number of shields in the store. It was bigger as it was not corrected for military and material shields that were already removed.